[CR-578] Move convert_rosbag2 script to slamcore_utils repo
Move convert_rosbag2 script from dataset-converter to the slamcore_utils
repo (atm the private one)

The moved files include:

* Main script - `convert_rosbag2` - renamed to `slamcore-convert-rosbag2` when
  installed via poetry or via package - for consistency with existing
  convention
* Helper types and functions such as `measurement_type`
* Main test file and test data files.

This is under a pip/poetry extra  called `ros2` so that users that don't
need it don't have to install extra dependencies such as `pillow` or
`lark` (latter is has to be present for workign with `rosbag2` and
`ros2` sourced installations). - Use `poetry install --extras ros2` on
the CLI to install it.

I also introduce a plugin mechanism so that the user of the tool can convert
arbitrary topics to new directories under the output directory structure. This
can be useful so that we don't hardcode but still convert the ROS 2 wrapper
output topics such as the TrackingStatus or the SLAMEvent.

To do this, I'm adding a slamcore_ros2_wrapper_topic_converters.py in the ROS 2
Wrapper. The user can specify the path to that using the `-p/--plugins`
argument, for example:

```
slamcore-convert-rosbag2 \
  -b trimmed_mh_01/trimmed_mh_01_0.db3 \
  -o output \
  -c cfg.json \
  -p ~/slamcore/slamcore_ros2/src/slamcore_ros2_wrapper/slamcore_extras/conversion_plugin/slamcore_ros2_wrapper_topic_converters.py -vv
```

With an example `cfg.json` file with the following contents (same as in earlier iterations of the convert_rosbag2 script):

```json
{
  "ir0": {
    "topic": "/slamcore/visible_0/image_raw"
  },
  "ir1": {
    "topic": "/slamcore/visible_1/image_raw"
  },

  "tracking_status0": {
    "topic": "/slamcore/metadata/tracking_status"
  },
  "slam_event0": {
    "topic": "/slamcore/metadata/slam_event"
  },
  "slam_pose0": {
    "topic": "/slamcore/pose"
  },
  "smooth_pose0": {
    "topic": "/slamcore/odom"
  }
}
```