[CR-964] Limit images queue memory to 2000 images
Limit the queue size for reading/writing images. Deserialization +
putting to the queue is much faster than consuming elements from the
Queue which would lead to unbounded mem usage.

I'm limiting the queue to a max of 2000 objects. Assuming uncompressed
grayscale images of `848x480` this would limit the used memory to <1GB