[CR-1204] Enable converting nav_msgs/Path messages
In arroyo-willow the customer also writes the optimized trajectory to
the bags they sent us for investigation / further analysis. This is
captured as a single `nav_msgs/msg/Path` message.

This commit adds a conversion plugin in `nav_path` which converts the
said message into a CSV file of stamped poses so that we can later use
this during ground-truth alignment.

Also includes:

* Minor formatting / text-width changes.
* Use `raise X from e` when inside an exception handler.